backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
psycopg2-binary
alembic
gitpython
sentence-transformers
faiss-cpu
python-dotenv
loguru